Rename 'Tokenizer' to 'HtmlTokenizer' and refactor the relevant code
The 'Tokenizer' class is renamed to 'HtmlTokenizer' for better context clarity. Similarly, 'TokenizerState' enum is updated to 'HtmlTokenizerState'.
Reason being, specificity of the class and the enum to be used for HTML parsing is now made clear in the name.
Also, added 'SelfClosing' attribute to the 'StartTagToken' class to handle self-closing HTML tags.
This change will improve the identification of self-closing Tags in the HTML parsing.
Many state handling functions are also added into 'HtmlTokenizer' class.
This detailed implementation ensures complete and correct parsing as specified by the HTML standard.